+++ src/js/components/parallax.js
@@ -1,5 +1,5 @@
 import Parallax from '../mixin/parallax';
-import {clamp, css, parent, query, scrolledOver} from 'uikit-util';
+import {clamp, css, parent, query, scrolledOver, toPx} from 'uikit-util';
 
 export default {
 
@@ -8,19 +8,31 @@ export default {
     props: {
         target: String,
         viewport: Number,
-        easing: Number
+        easing: Number,
+        start: String,
+        end: String
     },
 
     data: {
         target: false,
         viewport: 1,
-        easing: 1
+        easing: 1,
+        start: 0,
+        end: 0
     },
 
     computed: {
 
         target({target}, $el) {
             return getOffsetElement(target && query(target, $el) || $el);
+        },
+
+        start({start}, $el) {
+            return parseCalc(start, $el);
+        },
+
+        end({end}, $el) {
+            return parseCalc(end, $el);
         }
 
     },
@@ -38,11 +50,16 @@ export default {
             }
 
             const prev = percent;
-            percent = ease(clamp(scrolledOver(this.target), 0, this.viewport) / this.viewport, this.easing);
+            percent = ease(
+                clamp(
+                    scrolledOver(this.target, this.start, this.end), 0, this.viewport
+                ) / this.viewport,
+                this.easing
+            );
 
             return {
                 percent,
-                style: prev !== percent ? this.getCss(percent) : false
+                style: prev === percent ? false : this.getCss(percent)
             };
         },
 
@@ -61,6 +78,19 @@ export default {
     }
 
 };
+
+const calcRe = /-?\d+(?:\.\d+)?(?:v[wh]|%|px)?/g;
+function parseCalc(calc, el) {
+    let match;
+    let result = 0;
+    calc = calc.toString().replace(/\s/g, '');
+    calcRe.lastIndex = 0;
+    while ((match = calcRe.exec(calc)) !== null) {
+        result += toPx(match, 'height', el, true);
+    }
+
+    return result;
+}
 
 function ease(percent, easing) {
     return clamp(percent * (1 - (easing - easing * percent)));